LayoutTests/platform/mac/accessibility/img-fallsback-to-title-expected.txt

     test
test
This tests that images will fallback to using the title attribute if no other descriptive text is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Image1:
	AXTitle:
	AXDescription: test1
	AXHelp:
Image2:
	AXTitle:
	AXDescription: alt
	AXHelp: test2
Image3:
	AXTitle:
	AXDescription: test3
	AXHelp:
Image4:
	AXTitle:
	AXDescription: alt
	AXHelp: test4
PASS imagesGroup.childAtIndex(0).childrenCount is 2
PASS successfullyParsed is true

TEST COMPLETE
